refactor: extract shared model dataclasses into common module
Move ModelConfiguration, ModelSamples, and ModelFitContext from
individual model files into src/vocab_growth/models/common.py.

These three dataclasses were duplicated identically across all four
model modules (vg01-vg04). The shared module eliminates ~240 lines
of duplication per model file and provides a single source of truth.
